[Refactor][Kernel] Consolidate attention kernel directories into kernels/attention/
Merge flash_attn/, flash_decode/, deepseek_mla/ (attention parts), and
deepseek_nsa/ into a single kernels/attention/ subpackage. Extract
non-attention kernels (fp8_lighting_indexer, fp8_quant, topk_selector)
to kernels/ root. Update all imports and manifest paths.

Part 1 of #924.
