+++ tileops/kernels/attention/__init__.py
@@ -1,0 +1,60 @@
+from .deepseek_dsa_decode import SparseMlaKernel
+from .deepseek_mla_decode import MlaDecodeKernel, MlaDecodeWsKernel
+from .deepseek_nsa_cmp_fwd import NSACmpFwdVarlenKernel
+from .deepseek_nsa_fwd import NSAFwdVarlenKernel
+from .deepseek_nsa_mean_pooling_fwd import MeanPoolingFwdKernel
+from .deepseek_nsa_topk import NSATopkVarlenKernel
+from .gqa_bwd import (
+    FlashAttnBwdPostprocessKernel,
+    FlashAttnBwdPreprocessKernel,
+    GqaBwdKernel,
+    GqaBwdWgmmaPipelinedKernel,
+    MhaBwdKernel,
+    MhaBwdWgmmaPipelinedKernel,
+)
+from .gqa_decode import GqaDecodeKernel
+from .gqa_decode_paged import GqaDecodePagedKernel
+from .gqa_fwd import (
+    GqaFwdKernel,
+    GqaFwdWgmmaPipelinedKernel,
+    MhaFwdKernel,
+    MhaFwdWgmmaPipelinedKernel,
+)
+from .gqa_sliding_window_fwd import (
+    GqaSlidingWindowFwdKernel,
+    GqaSlidingWindowFwdWgmmaPipelinedKernel,
+)
+from .gqa_sliding_window_varlen_fwd import (
+    GqaSlidingWindowVarlenFwdKernel,
+    GqaSlidingWindowVarlenFwdWgmmaPipelinedKernel,
+)
+from .mha_decode import MhaDecodeKernel
+from .mha_decode_paged import MhaDecodePagedKernel
+
+__all__ = [
+    "FlashAttnBwdPostprocessKernel",
+    "FlashAttnBwdPreprocessKernel",
+    "GqaBwdKernel",
+    "GqaBwdWgmmaPipelinedKernel",
+    "GqaDecodeKernel",
+    "GqaDecodePagedKernel",
+    "GqaFwdKernel",
+    "GqaFwdWgmmaPipelinedKernel",
+    "GqaSlidingWindowFwdKernel",
+    "GqaSlidingWindowFwdWgmmaPipelinedKernel",
+    "GqaSlidingWindowVarlenFwdKernel",
+    "GqaSlidingWindowVarlenFwdWgmmaPipelinedKernel",
+    "MeanPoolingFwdKernel",
+    "MhaBwdKernel",
+    "MhaBwdWgmmaPipelinedKernel",
+    "MhaDecodeKernel",
+    "MhaDecodePagedKernel",
+    "MhaFwdKernel",
+    "MhaFwdWgmmaPipelinedKernel",
+    "MlaDecodeKernel",
+    "MlaDecodeWsKernel",
+    "NSACmpFwdVarlenKernel",
+    "NSAFwdVarlenKernel",
+    "NSATopkVarlenKernel",
+    "SparseMlaKernel",
+]
