+++ tileops/kernels/attention/deepseek_nsa_mean_pooling_fwd.py
