--- tileops/kernels/deepseek_mla/__init__.py
@@ -1,14 +1,0 @@
-from .deepseek_dsa_decode import SparseMlaKernel
-from .deepseek_mla_decode import MlaDecodeKernel, MlaDecodeWsKernel
-from .fp8_lighting_indexer import FP8LightingIndexerKernel
-from .fp8_quant import FP8QuantKernel
-from .topk_selector import TopkSelectorKernel
-
-__all__ = [
-    "FP8LightingIndexerKernel",
-    "FP8QuantKernel",
-    "SparseMlaKernel",
-    "TopkSelectorKernel",
-    "MlaDecodeKernel",
-    "MlaDecodeWsKernel",
-]
