--- tileops/kernels/deepseek_nsa/__init__.py
@@ -1,23 +1,0 @@
-from .gqa_sliding_window_fwd import (
-    GqaSlidingWindowFwdKernel,
-    GqaSlidingWindowFwdWgmmaPipelinedKernel,
-)
-from .gqa_sliding_window_varlen_fwd import (
-    GqaSlidingWindowVarlenFwdKernel,
-    GqaSlidingWindowVarlenFwdWgmmaPipelinedKernel,
-)
-from .mean_pooling_fwd import MeanPoolingFwdKernel
-from .nsa_cmp_fwd import NSACmpFwdVarlenKernel
-from .nsa_fwd import NSAFwdVarlenKernel
-from .nsa_topk import NSATopkVarlenKernel
-
-__all__ = [
-    "MeanPoolingFwdKernel",
-    "NSAFwdVarlenKernel",
-    "NSATopkVarlenKernel",
-    "NSACmpFwdVarlenKernel",
-    "GqaSlidingWindowFwdKernel",
-    "GqaSlidingWindowFwdWgmmaPipelinedKernel",
-    "GqaSlidingWindowVarlenFwdKernel",
-    "GqaSlidingWindowVarlenFwdWgmmaPipelinedKernel",
-]
