--- tileops/kernels/flash_attn/__init__.py
@@ -1,22 +1,0 @@
-from .bwd import (
-    FlashAttnBwdPostprocessKernel,
-    FlashAttnBwdPreprocessKernel,
-    GqaBwdKernel,
-    GqaBwdWgmmaPipelinedKernel,
-    MhaBwdKernel,
-    MhaBwdWgmmaPipelinedKernel,
-)
-from .fwd import GqaFwdKernel, GqaFwdWgmmaPipelinedKernel, MhaFwdKernel, MhaFwdWgmmaPipelinedKernel
-
-__all__ = [
-    "FlashAttnBwdPostprocessKernel",
-    "FlashAttnBwdPreprocessKernel",
-    "GqaBwdKernel",
-    "GqaBwdWgmmaPipelinedKernel",
-    "GqaFwdKernel",
-    "GqaFwdWgmmaPipelinedKernel",
-    "MhaBwdKernel",
-    "MhaBwdWgmmaPipelinedKernel",
-    "MhaFwdKernel",
-    "MhaFwdWgmmaPipelinedKernel",
-]
