--- tileops/kernels/flash_decode/__init__.py
@@ -1,11 +1,0 @@
-from .gqa_decode import GqaDecodeKernel
-from .gqa_decode_paged import GqaDecodePagedKernel
-from .mha_decode import MhaDecodeKernel
-from .mha_decode_paged import MhaDecodePagedKernel
-
-__all__ = [
-    "GqaDecodeKernel",
-    "GqaDecodePagedKernel",
-    "MhaDecodeKernel",
-    "MhaDecodePagedKernel",
-]
