+++ tileops/ops/attention/deepseek_dsa_decode.py
@@ -2,7 +2,7 @@ from typing import Dict, Optional
 
 import torch
 
-from tileops.kernels.deepseek_mla import SparseMlaKernel
+from tileops.kernels.attention import SparseMlaKernel
 from tileops.kernels.kernel import Kernel
 
 from ..op import Op
