+++ tileops/ops/attention/mha_decode_paged.py
@@ -2,7 +2,7 @@ from typing import Dict, Optional
 
 import torch
 
-from tileops.kernels.flash_decode.mha_decode_paged import MhaDecodePagedKernel
+from tileops.kernels.attention.mha_decode_paged import MhaDecodePagedKernel
 from tileops.kernels.kernel import Kernel
 
 from ..op import Op
